Fix second bug in random test. When Random number generated it can contain \0 which will affect string length.
Reinstate file_store function. This version automatically creates a backup of an existing file then stores the file.

This functions contains a bug. Without the backup the function will just write new data to the beginning of an existing file, thus overwriting any data that already exists. It is maintained for compatibility with the existing software but will be replaced.